AG's News Topic Classification Dataset

The AG's news topic classification dataset is constructed by choosing 4 largest classes from the original corpus. Each class contains 30,000 training samples and 1,900 testing samples. The total number of training samples is 120,000 and testing 7,600. Version 3, Updated 09/09/2015

dataset_ag_news(
  dir = NULL,
  split = c("train", "test"),
  delete = FALSE,
  return_path = FALSE,
  clean = FALSE
)

Arguments

dir
Character, path to directory where data will be stored. If NULL, user_cache_dir will be used to determine path.

split
Character. Return training ("train") data or testing ("test") data. Defaults to "train".

delete
Logical, set TRUE to delete dataset.

return_path
Logical, set TRUE to return the path of the dataset.

clean
Logical, set TRUE to remove intermediate files. This can greatly reduce the size. Defaults to FALSE.

Source

http://www.di.unipi.it/~gulli/AG_corpus_of_news_articles.html

https://github.com/mhjabreel/CharCnn_Keras/tree/master/data/ag_news_csv

Value

A tibble with 120,000 or 30,000 rows for "train" and "test" respectively and 3 variables:

class

Character, denoting new class

title

Character, title of article

description

Character, description of article

Details

The classes in this dataset are

World

Sports

Business

Sci/Tech

See also

Examples

# \donttest{
dataset_ag_news()

#> Do you want to download:
#>  Name: AG News 
#>  URL: https://www.di.unipi.it/~gulli/AG_corpus_of_news_articles.html 
#>  License: You are encouraged to download this corpus for any non-commercial use. 
#>  Size: 64.4 MB (cleaned 33.9 MB) 
#>  Download mechanism: https 

#> Error in menu(choices = c("Yes", "No"), title = title): menu() cannot be used non-interactively

# Custom directory
dataset_ag_news(dir = "data/")

#> Do you want to download:
#>  Name: AG News 
#>  URL: https://www.di.unipi.it/~gulli/AG_corpus_of_news_articles.html 
#>  License: You are encouraged to download this corpus for any non-commercial use. 
#>  Size: 64.4 MB (cleaned 33.9 MB) 
#>  Download mechanism: https 

#> Error in menu(choices = c("Yes", "No"), title = title): menu() cannot be used non-interactively

# Deleting dataset
dataset_ag_news(delete = TRUE)

#> Error: [ENOENT] Failed to search directory '/home/travis/.cache/textdata/ag_news': no such file or directory

# Returning filepath of data
dataset_ag_news(return_path = TRUE)

#> /home/travis/.cache/textdata/ag_news

# Access both training and testing dataset
train <- dataset_ag_news(split = "train")

#> Do you want to download:
#>  Name: AG News 
#>  URL: https://www.di.unipi.it/~gulli/AG_corpus_of_news_articles.html 
#>  License: You are encouraged to download this corpus for any non-commercial use. 
#>  Size: 64.4 MB (cleaned 33.9 MB) 
#>  Download mechanism: https 

#> Error in menu(choices = c("Yes", "No"), title = title): menu() cannot be used non-interactively

test <- dataset_ag_news(split = "test")

#> Do you want to download:
#>  Name: AG News 
#>  URL: https://www.di.unipi.it/~gulli/AG_corpus_of_news_articles.html 
#>  License: You are encouraged to download this corpus for any non-commercial use. 
#>  Size: 64.4 MB (cleaned 33.9 MB) 
#>  Download mechanism: https 

#> Error in menu(choices = c("Yes", "No"), title = title): menu() cannot be used non-interactively

# }