AFINN-111 dataset

AFINN is a lexicon of English words rated for valence with an integer between minus five (negative) and plus five (positive). The words have been manually labeled by Finn Årup Nielsen in 2009-2011.

lexicon_afinn(dir = NULL, delete = FALSE, return_path = FALSE, clean = FALSE)

Arguments

dir
Character, path to directory where data will be stored. If NULL, user_cache_dir will be used to determine path.

delete
Logical, set TRUE to delete dataset.

return_path
Logical, set TRUE to return the path of the dataset.

clean
Logical, set TRUE to remove intermediate files. This can greatly reduce the size. Defaults to FALSE.

Value

A tibble with 2,477 rows and 2 variables:

word

An English word

sentiment

Indicator for sentiment: integer between -5 and +5

Details

This dataset is the newest version with 2477 words and phrases.

Citation info:

This dataset was published in Finn Ärup Nielsen (2011), ``A new ANEW: Evaluation of a word list for sentiment analysis in microblogs'', Proceedings of the ESWC2011 Workshop on 'Making Sense of Microposts': Big things come in small packages (2011) 93-98.

article{nielsen11,
author = {Finn Äruprup Nielsen},
title = {A new ANEW: Evaluation of a word list for sentiment analysis in microblogs},
journal = {CoRR},
volume = {abs/1103.2903},
year = {2011},
url = {http://arxiv.org/abs/1103.2903},
archivePrefix = {arXiv},
eprint = {1103.2903},
biburl = {https://dblp.org/rec/bib/journals/corr/abs-1103-2903},
bibsource = {dblp computer science bibliography, https://dblp.org}
}

See also

Examples

# \donttest{
lexicon_afinn()

#> Do you want to download:
#>  Name: AFINN-111 
#>  URL: http://www2.imm.dtu.dk/pubdb/views/publication_details.php?id=6010 
#>  License: Open Database License (ODbL) v1.0 
#>  Size: 78 KB (cleaned 59 KB) 
#>  Download mechanism: https 

#> Error in menu(choices = c("Yes", "No"), title = title): menu() cannot be used non-interactively

# Custom directory
lexicon_afinn(dir = "data/")

#> Do you want to download:
#>  Name: AFINN-111 
#>  URL: http://www2.imm.dtu.dk/pubdb/views/publication_details.php?id=6010 
#>  License: Open Database License (ODbL) v1.0 
#>  Size: 78 KB (cleaned 59 KB) 
#>  Download mechanism: https 

#> Error in menu(choices = c("Yes", "No"), title = title): menu() cannot be used non-interactively

# Deleting dataset
lexicon_afinn(delete = TRUE)

#> Error: [ENOENT] Failed to search directory '/home/travis/.cache/textdata/afinn': no such file or directory

# Returning filepath of data
lexicon_afinn(return_path = TRUE)

#> /home/travis/.cache/textdata/afinn

# }