The NRC Valence, Arousal, and Dominance Lexicon

The NRC Valence, Arousal, and Dominance (VAD) Lexicon includes a list of more than 20,000 English words and their valence, arousal, and dominance scores. For a given word and a dimension (V/A/D), the scores range from 0 (lowest V/A/D) to 1 (highest V/A/D). The lexicon with its fine-grained real- valued scores was created by manual annotation using best--worst scaling. The lexicon is markedly larger than any of the existing VAD lexicons. We also show that the ratings obtained are substantially more reliable than those in existing lexicons.

lexicon_nrc_vad(dir = NULL, delete = FALSE, return_path = FALSE,
  clean = FALSE)

Arguments

dir
Character, path to directory where data will be stored. If NULL, user_cache_dir will be used to determine path.

delete
Logical, set TRUE to delete dataset.

return_path
Logical, set TRUE to return the path of the dataset.

clean
Logical, set TRUE to remove intermediate files. This can greatly reduce the size. Defaults to FALSE.

Source

https://saifmohammad.com/WebPages/nrc-vad.html

Value

A tibble with 20.007 rows and 4 variables:

word

An English word

Valence

valence score of the word

Arousal

arousal score of the word

Dominance

dominance score of the word

Details

License required for commercial use. Please contact Saif M. Mohammad (saif.mohammad@nrc-cnrc.gc.ca).

Citation info:

Details of the NRC VAD Lexicon are available in this paper:

Obtaining Reliable Human Ratings of Valence, Arousal, and Dominance for 20,000 English Words. Saif M. Mohammad. In Proceedings of the 56th Annual Meeting of the Association for Computational Linguistics, Melbourne, Australia, July 2018.

inproceedings{vad-acl2018,
title={Obtaining Reliable Human Ratings of Valence, Arousal, and Dominance for 20,000 English Words},
author={Mohammad, Saif M.},
booktitle={Proceedings of The Annual Conference of the Association for Computational Linguistics (ACL)},
year={2018},
address={Melbourne, Australia}
}

See also

Examples

lexicon_nrc_vad()

#> Do you want to download:
#>  Name: The NRC Valence, Arousal, and Dominance Lexicon 
#>  URL: https://saifmohammad.com/WebPages/nrc-vad.html 
#>  License: License required for commercial use. Please contact Saif M. Mohammad (saif.mohammad@nrc-cnrc.gc.ca). 
#>  Size: 150.8 MB (cleaned 792 KB) 
#>  Download mechanism: http 
#>  Citation info:
#> 
#> inproceedings{vad-acl2018,
#> title={Obtaining Reliable Human Ratings of Valence, Arousal, and Dominance for 20,000 English Words},
#> author={Mohammad, Saif M.},
#> booktitle={Proceedings of The Annual Conference of the Association for Computational Linguistics (ACL)},
#> year={2018},
#> address={Melbourne, Australia}
#> }
#> 
#> If you use this lexicon, then please cite it. 

#> Error in menu(choices = c("Yes", "No"), title = title): menu() cannot be used non-interactively

# Custom directory
lexicon_nrc_vad(dir = "data/")

#> Do you want to download:
#>  Name: The NRC Valence, Arousal, and Dominance Lexicon 
#>  URL: https://saifmohammad.com/WebPages/nrc-vad.html 
#>  License: License required for commercial use. Please contact Saif M. Mohammad (saif.mohammad@nrc-cnrc.gc.ca). 
#>  Size: 150.8 MB (cleaned 792 KB) 
#>  Download mechanism: http 
#>  Citation info:
#> 
#> inproceedings{vad-acl2018,
#> title={Obtaining Reliable Human Ratings of Valence, Arousal, and Dominance for 20,000 English Words},
#> author={Mohammad, Saif M.},
#> booktitle={Proceedings of The Annual Conference of the Association for Computational Linguistics (ACL)},
#> year={2018},
#> address={Melbourne, Australia}
#> }
#> 
#> If you use this lexicon, then please cite it. 

#> Error in menu(choices = c("Yes", "No"), title = title): menu() cannot be used non-interactively

# Deleting dataset
lexicon_nrc_vad(delete = TRUE)

#> Error: [ENOENT] Failed to search directory '/home/travis/.cache/textdata/nrc_vad': no such file or directory

# Returning filepath of data
lexicon_nrc_vad(return_path = TRUE)

#> /home/travis/.cache/textdata/nrc_vad